Add support for protobufs in puree-android.
This change is fairly large and opinionated. Puree Android previously worked
entirely with JSON via GSON's JsonObject and family. In order to add generic
protobuf support, I essentially split Puree in two so that it can accept
either JSON log entries OR protobuf log entries, even in the same application.
When registering a log class with PureeConfiguration, you need to specify how
the log should be stored. Log classes which extend PureeLog will use the JSON
logging pathway, and classes which extend MessageLite (the lite version of
protobuf) will automatically use the Protobuf logging pathway.

At the storage level, I added a "format" column to the table which keeps track
of whether a log entry is JSON or protobuf. To simplify things, for now protobuf
is serialized to bytes and then encoded as a base64 string for storage in the
same table. We can probably split into multiple tables in the future, but this
is the simplest solution and is fast enough for our needs since buffered logs
get sent on background threads anyway.

In order to support differences between databases, I simply changed the database
name itself that's used for the sqlite database rather than changing version numbers.
The reason is the following: It's possible that an app might use the original version
of Puree, then upgrade to this one which has a different database format. The upgrade
could work properly with onUpgrade and incrementing the version number, but if for
some reason there was a rollback to a version of Puree from the main repository
rather than this fork, then you could potentially find yourself in a situation where
there was an incompatibility because onDowngrade is not implemented in the main repo.
A simple solution is just changing the database name itself, because in the event of
a rollback to the version in the main repo, that version uses a different database
name and onDowngrade would never be fired.

In order to support protobuf in a generic way, the storage system simply stores the
serialized bytes from the protobuf MessageLite instance, but does not store the class
itself, so the storage system cannot repopulate the object. It's up to the output
plugins to know what type is stored with each log entry. See OutBufferedProtobufLogcat.java
for an example of repopulating a protobuf from the buffer and then wrapping it in another
protobuf.

I upgraded the demo to show how to use protobufs for logging. The protobuf itself
is defined in demo/src/main/proto/example.proto and is a bare minimum protobuf.
The demo also shows how to properly set up gradle on apps in order to compile
protobufs for usage in the app, if that's your thing.

I also fixed a bunch of bugs I found along the way.

TESTED=original unit tests and demo.